Add basic validation

+++ api/YourNeighbour.Application/Features/CategoryDefinitions/Commands/DeleteCategoryDefinition/DeleteCategoryDefinitionHandler.cs
@@ -4,8 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using FluentValidation;
 using YourNeighbour.Application.Abstractions;
 using YourNeighbour.Data.Interfaces.Repositories;
+using YourNeighbour.Domain.Entities.Definitions;
 
 namespace YourNeighbour.Application.Features.CategoryDefinitions.Commands.DeleteCategoryDefinition
 {
@@ -19,6 +21,10 @@ namespace YourNeighbour.Application.Features.CategoryDefinitions.Commands.Delete
         }
         public async Task<bool> Handle(DeleteCategoryDefinitionCommand request, CancellationToken cancellationToken)
         {
+            if (await categoryDefinitionRepository.IsBasicById(request.Id))
+                throw new ValidationException("Nie można usunąć podstawowej definicji.");
+            if (await categoryDefinitionRepository.HasAnyCategoryById(request.Id))
+                throw new ValidationException("Definicja ma przypisane kategorie.");
             return await categoryDefinitionRepository.Delete(request.Id);
         }
     }
